Move exception related files into appropriate modules Add MathesarErrorMixin to serializers in order to change the validation errors into correct error format that conforms to the spec
