+++ mathesar/api/serializers/records.py
@@ -1,12 +1,14 @@
 from rest_framework import serializers
 
+from mathesar.api.exceptions.mixins import MathesarErrorMessageMixin
 
-class RecordListParameterSerializer(serializers.Serializer):
+
+class RecordListParameterSerializer(MathesarErrorMessageMixin, serializers.Serializer):
     filters = serializers.JSONField(required=False, default=[])
     order_by = serializers.JSONField(required=False, default=[])
     group_count_by = serializers.JSONField(required=False, default=[])
 
 
-class RecordSerializer(serializers.BaseSerializer):
+class RecordSerializer(MathesarErrorMessageMixin, serializers.BaseSerializer):
     def to_representation(self, instance):
         return instance._asdict()
