+++ mathesar/errors.py
@@ -1,0 +1,4 @@
+
+
+class InvalidTableError(Exception):
+    pass
